refactor: Fix comments
