+++ packages/tangle/tangle.go
@@ -13,6 +13,15 @@ import (
 )
 
 const (
+	// PrefixMessage defines the storage prefix for message.
+	PrefixMessage byte = iota
+	// PrefixMessageMetadata defines the storage prefix for message metadata.
+	PrefixMessageMetadata
+	// PrefixApprovers defines the storage prefix for approvers.
+	PrefixApprovers
+	// PrefixMissingMessage defines the storage prefix for missing message.
+	PrefixMissingMessage
+
 	cacheTime = 20 * time.Second
 )
 
@@ -30,8 +39,8 @@ type Tangle struct {
 	shutdown               chan struct{}
 }
 
-// NewTangle creates a new Tangle.
-func NewTangle(store kvstore.KVStore) (result *Tangle) {
+// New creates a new Tangle.
+func New(store kvstore.KVStore) (result *Tangle) {
 	osFactory := objectstorage.NewFactory(store, storageprefix.MessageLayer)
 
 	result = &Tangle{
